requirements.txt
alpaca-py>=0.21.0
python-dotenv>=1.0.0
pandas>=2.0.0
